rename and separate file
